t_bot/pyproject.toml
[tool.poetry]
name = "zkauth-solana-tgbot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "zkauth_solana_tgbot"}]

[tool.poetry.dependencies]
python = "^3.9"
python-telegram-bot = "^20.1"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
